requirements.txt
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
decorator==4.4.2
Flask==1.1.2
future==0.18.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycparser==2.20
python-decouple==3.3
python-telegram-bot==12.7
requests==2.23.0
six==1.14.0
tornado==6.0.4
urllib3==1.25.9
Werkzeug==1.0.1
